Refactor to inheritance approach to maintain sanity
